requirements.txt
python-telegram-bot==13.7
Flask==2.0.1
python-dotenv==0.19.1